+++ src/collective/contact/plonegroup/utils.py
@@ -8,6 +8,7 @@ from collective.contact.plonegroup.config import set_registry_functions
 from collective.contact.plonegroup.config import set_registry_organizations
 from html import escape
 from imio.helpers.cache import get_users_in_plone_groups
+from imio.helpers.content import find
 from imio.helpers.content import get_user_fullname
 from imio.helpers.content import uuidsToObjects
 from imio.helpers.content import uuidToObject
@@ -22,6 +23,25 @@ from zope.globalrequest import getRequest
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
+
+
+def get_person_from_userid(userid, context=None, depth=None, unrestricted=False, objects=True):
+    """Return person from userid.
+
+    :param userid: mandatory string
+    :param context: context to search (default None)
+    :param depth: depth to search (default None)
+    :param unrestricted: search unrestrictedly (default False)
+    :param objects: return objects list (default True) or brains (False)
+    :return: object or brains list
+    """
+    brains = find(context=None, depth=None, unrestricted=False, userid=userid)
+    if objects:
+        if unrestricted:
+            return [br._unrestrictedGetObject() for br in brains]
+        else:
+            return [br.getObject() for br in brains]
+    return brains
 
 
 def organizations_with_suffixes(groups, suffixes, group_as_str=False):