+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - Added `behaviors.IPlonegroupUserLink` with userid and primary_organization fields.
   [sgeulette]
+- Added `utils.get_person_from_userid`
+  [sgeulette]
 
 1.47 (2023-10-19)
 -----------------